+++ auth/src/repository/sqlite/sqlite_auth_account.rs
@@ -36,7 +36,7 @@ impl AuthAccountRepository for SqliteAuthAccountRepository {
             limit ?
         "#,
         )
-        .bind(start_user_id as i64)
+        .bind(start_user_id)
         .bind(status.as_ref())
         .bind(limit as i64)
         .fetch_all(tx)
